add support for fetching author's url in pull requests
added support for fetching the author's url in pull requests in order to properly format the contributors list in the changelog. this was accomplished by updating the pullrequest, author, and query.graphql structs, adding the url field to the author struct and the query, and calling the field in the map_pull_request function. finally, a new format_contributors_to_md function was added to the main.rs file in order to properly format the contributors in the changelog.