update main.rs to remove unnecessary result type
updates the `main.rs` file to remove the unnecessary `result` type and replace it with a `main()` function that returns nothing.

the changes made to the file are as follows:

- changed the `main()` function signature from `result<(), box<dyn std::error::error>>` to `main()`
- removed the call to `ok` at the end of the function

in addition to the changes in `main.rs`, this commit also updates the `github_graphql/mod.rs` file to add the `reqwest::header` module back in.