adding support for github releases
adds support for github releases in the program. it now takes an additional argument `--github_token` which will be used to authenticate with the github api, and a new struct `args` has been added to the `main.rs` file to accomodate for these changes.

```diff
diff --git a/src/main.rs b/src/main.rs
index acf05070b878..e383e4531740 100644
--- a/src/main.rs
+++ b/src/main.rs
@@ -10,17 +10,12 @@ mod github_graphql;
 #[derive(parser, debug)]
 #[command(author, version, about, long_about = none)]
 struct args {
-  /// name of the person to greet
   #[arg(short, long)]
   release: string,
-
   #[arg(short, long)]
   owner: string,
-
   #[arg(short, long)]
   project: string,
-
-  /// github token
   #[arg(short, long)]
   github_token: string,
 }
```

this commit adds support for github releases in the program. it adds a new argument `--github_token` which will be used to authenticate with the github api, and a new struct `args` has been added to the `main.rs` file to accomodate for these changes. the changes to the file can be seen in the diff below.