add logging to the app
adds logging to the app by adding the `logger` module in `src/` and including it in the `main.rs` file.

the `logger` module contains the `init_logger` function, which sets up the environment logger and sets the log level to `info` (`log::levelfilter::info`). this function is then called in `main.rs` when the application starts.

the changes to the code can be seen in the diff:
- added `mod logger` to `src/main.rs`
- added `src/logger/mod.rs` with the `init_logger` function
- removed `builder` from `main.rs` and replaced it with a call to `logger::init_logger()`