add submodules to decompose the structure
adds a template for a changelog in the src/templates directory. this template is written using the askama crate and provides structure for the changelog.

this commit also updates the main.rs file to use the newly created template. this includes removing the unused chrono::naivedate import and adding a mod declaration for the templates directory.