updated readme to remove deprecated options
removes the deprecated options from the readme to better reflect the current features of the project.

the main change was made to the following section:

```
to see all values suported, look at the `changelog.config.js` file at the root of this repo.
```

this line was removed in this commit as the options are no longer supported.